pass args to generate the meta merkle tree

+++ tip-router-operator-cli/src/lib.rs
@@ -77,7 +77,7 @@ pub fn get_merkle_root(
     tip_distribution_program_id: &Pubkey,
     out_path: &str,
     tip_payment_program_id: &Pubkey,
-    protocol_fee_bps: u16,
+    protocol_fee_bps: u64,
 ) -> std::result::Result<MetaMerkleTree, MerkleRootError> {
     // Get stake meta collection
     let stake_meta_collection = stake_meta_generator::generate_stake_meta(
